renamed structure, memberbox check in
