+++ src/api/mock.js
@@ -22,6 +22,7 @@ export const mock = [
           first_name: 'dfdfs',
           guest_image: 'https://picsum.photos/id/2/200/200',
           id: 2,
+          isCheckedIn: 'false',
           last_name: 'manchuri',
           membership: 'guest',
           phone_number: '45352',
@@ -30,6 +31,7 @@ export const mock = [
           first_name: 'patrick',
           guest_image: 'https://picsum.photos/id/1/200/200',
           id: 1,
+          isCheckedIn: 'false',
           last_name: 'starr',
           membership: 'guest',
           phone_number: '920000',
